MaterialSystem: increase static limit of streaming textures
The static array of TextureStreamMgr_s::streamingTextures[] was insreased from 16384 to 32768 using the new array s_streamingTextureHandles. A pointer to this new array is now installed in the old array and the assembly code of the game dereferences it to utilize our new array.
This change did not incur a performance penalty, as the newly patched in dereference happens outside the hot code. The new array has been tested on new paks totalling a total streaming texture count > 16384 and the game engine was now able to handle this perfectly.