Thirdparty: add Boost C++ Libraries
Boost C++ Libraries v1.87 is needed to have char_conv in msvc 15.6 (Visual Studio 2017) as we have to retain ABI compatibility with the game executable, which was compiled with this compiler. This will also be used to replace std::regex with boost::regex as boost::regex is a lot faster than std::regex.
- Only char_conv and regex have been included from boost, with all their dependencies.