Tier0: rework signature caching class
The signature is now hashed using Murmur64, which will be used as the key to extract the RVA from the static file. Compression has been removed as this was no longer necessary (it barely compressed as we are storing the hash now). The buffers are now also written, and read using streams rather than a deep copy in memory.