Engine: optimize CClient::Authenticate()
Don't run formatting and claim parsing code if the token or signature is missing or empty, and use the already computed claim value size instead of doing a strlen() on it. Also separated the missing signature of the token into its own error, and added errors when either the token or its signature was empty.