Common: fix "changelevel" command handler when loaded in the lobby
The "map" command didn't work in mp_lobby, because the "map" command calls the original "changelevel" command handler internally if we are loaded in a map. If we are in mp_lobby, the code enforces additional checks on the local Origin and Nucleus token. These tokens are not required for local multiplayer play, only if you play on secure multiplayer server, and therefore this check is redundant. Hook it against our own "changelevel" command handler so that all internal calls get routed through ours instead. This patch also fixes the CreateServer() script function not working when being loaded in the lobby.