Client: thread abstracted script function RequestEULAContents()
The script function hangs the main thread on call since the request itself wasn't threaded. Request is now threaded and results are pushed back to the main thread, which will be routed to the newly created script function UICodeCallback_OnEULARequestCompleted(). This completely eliminates the hanging issue, declaring this function fully implemented and finished.