Recast: make skipping off-mesh connection optional
Make skipping off-mesh connections optional in dtNavMeshQuery::raycast(), because there are situations where we do want to have this include off-mesh connection when we perform a raycast (i.e. player or NPC is at the start of the zipline rope, or traversing a door, and the other NPC wants to melee the player or NPC, but performs a raycast first before running the more CPU intensive logic).