Recast: fixup filter->traverseFilter() parameters
We should not give the poly and tile from which the link is originating from, but rather connecting to. This is not a bug because these parameters are unused, but they can be used when dtQueryFilter is made virtual and  dtQueryFilter::traverseFilter() is overriden. This change is to enforce consistency.