Recast: delay right vertex in case left lies outside the funnel
Since we changed the coordinate system, the left vertex must always be added first, but we need to check on the right vertex first and update the parameters if it lies outside the funnel. This fully fixes the string pulling algorithm. Also performed a major deduplication by collapsing the left and right segment adding logic.