Recast: refactor dtNavMesh::connectTraverseLinks
The large algorithm has now been broken up into smaller parts to make the code easier to read and understand without affecting performance. The various parameters are passed over using an internal state. The logic of the actual algorithm remains unchanged and the output is still identical to the code prior to this patch.