Server: rework NavMesh_GetNearestPos functions
Code now returns a script error if the NavMesh for a given hull wasn't found. Also used actual getter instead of using the raw objects, and improved the descriptions of the functions. Also removed the call to dtNavMeshQuery::findNearestPolyInBounds(), this function is deprecated as the behavior is the same as dtNavMeshQuery::findNearestPoly(). It will be removed in a future commit.