Engine: implement new shape renderers to DebugOverlay
The debug overlay was lacking a bunch of renderers as these were optimized away by the compiler. The SDK has rebuild the last 2 in commit 691ef3662e6bd6c0fc91d4ed0efa82ce9c3897d7 which allows for a full restoration on the DebugOverlay system. This also removes the need of the  `r_debug_overlay_invisible` hack as we can now draw boxes with faces, and boxes with - alpha should be wireframe only.