Engine: add proper support for capsule debug overlay
Implemented a proper interface for adding capsule overlays in CIVDebugOverlay (this cannot be virtual as we have to maintain backwards compatibility with the compiled modules of the game engine). This patch also makes use of the new capsule renderer introduced in commit 0bede8bc2ce48fe7f059869a1199cd3e4e277386.