Recast: remove hard neighbor limit from dtNavMesh::connectExtLinks
See recastnavigation/recastnavigation#756, except we do not enforce a limit at all since `dtNavMesh::addTile` no longer builds the links in our fork of the library. The entire linking process is offline and all the links are cached in the navmesh file, therefore we can use a dynamic array here.