ImGui: properly handle text alpha in color text logger
The text inside the color text logger did not handle alpha values correctly as setting the alpha through `ImGui::PushStyleVar()` will not have any effects on draw commands added directly to an ImDrawList. This patch passes the value directly into the color of the draw command to make the text fade away correctly as well.