ImGui:: fix flickering when render thread is ahead of time
This is a regression that was caused when the ImguiSystem was optimized to only run if there's something to render. The render thread can be a frame ahead and thus render the previous frame if the buffers hadn't been swapped yet, however the ImguiSystem did not account for this. Repeat the last frame once in case this occurs.