ImGui: improve console autocomplete suggest panel
Instead of calculating the size strictly from a hardcoded item rect height, use the cursor positions to determine the final height of the render and min it out on the cvar `con_autocomplete_window_height`. This allows for commands/convarsd with formatted help texts to make as much and efficient use of the available window space as possible instead of truncating it if everything combined still takes less height than `con_autocomplete_window_height`.