Engine: fix flickering transparent client overlays on low frame times
There is a bug in the engine where all overlays are assigned to 'g_nOverlayStage + 1' when the duration is set to 0. However, only server overlays should be set to 'g_nOverlayStage' (without the +1 increment). The +1 increment was added to soften the flickering effect on the client by rendering it for 2 frames which is a hack. The increment also causes the alpha to be multiplied as the overlay will always overlap on low frame times.
The proper fix here is to assign client overlays to the render tick count (similar to how splines are always set to this tick count). All server overlays originate from the server frame thread in this build so we can just check on ThreadInServerFrameThread() to determine which tick to use.