Engine: allow overriding depth-testing for every debug overlay
Previously, you weren't able to set the depth testing value for the sphere and swept box overlay, making the API inconsistent, confusing and limiting. This patch now adds the parameter to `IVDebugOverlay::AddSphereOverlay` and `IVDebugOverlay::AddSweptBoxOverlay`. Nothing in the engine was using the sphere overlay, so this didn't break anything. However, the swept box overlay was used by the `cast_hull` consone command, and the debug overlay for `automantle_debug`. These area;s of the engine have been patched in assembly and now take the new boolean parameter into account. They are tested and confirmed to work as intended.