Engine: make AddCapsuleOverlay virtual
The unknown overlay was never used, and its renderer also didn't exist (optimized out as it was never used in the engine executable). This is now replaced with AddCapsuleOverlay which allows us to make it virtual and accessible from anywhere (including plugin dll's in the future), since this will be used from the interface exposed by the engine.