Recast: fix crash during prune of post-unlinked off-mesh connection
If you build a NavMesh and remove a tile, and that tile happens to be the land tile of one or more off-mesh connection, it's polygon will still have the `DT_POLYFLAGS_JUMP_LINKED` flag causing a null pointer dereference in `dtUpdateNavMeshData()`. This function drops every off-mesh connection that doesn't have the `DT_POLYFLAGS_JUMP_LINKED`, but it fixes the polygon references within the land tile if it does have this flag. But since we removed this tile, `dtNavMesh::getTileAt()` will return NULL for the given tile location.
The fix is to update the polygon flags in `dtNavMesh::unconnectLinks()` when its an off-mesh connection.