Recast: fix singly linked list leak on off-mesh reconnect
Do not rebase the off-mesh connection if it was already based. I.e. if we remove the tile our off-mesh connection end point is landing on, and then re-add this tile, we only want to link this existing off-mesh connection back to this tile and maintain the old base link.