RTech: properly split RTech and ZStd pak encoding flags
Each encoding type should have its own flag. Previously the RTech encode flag was needed in the RPak header as we didn't had `Pak_SetupBuffersAndLoad()` rebuilt in the SDK. Now that we have it, we should set the header flags properly. Also renamed the defines to make them more explicit.