Engine: fix Nucleus ID leak caused by bug in CClient::Clear()
CClient::Clear() internally never nulls the Nucleus ID, this isn't an issue if this CClient instance gets reused for another human player as it gets overwritten by the new one. But when this CClient instance gets used for a bot, the Nucleus ID of the previous player will remain unchanged and therefore be reused which is incorrect behavior (g_pEngineServer->CreateFakeClient() doesn't set the Nucleus ID). Explicitly null it in CClient::Clear() to yield correct behavior.