RTech: fix rare deadlock caused by RPak files with ZStd compression
In commit 3f01894bfa063cc9f642906a74188f5a1bdb4787, a fix was applied making sure there's always an objective to stream in extra data to decode. However, `ZSTD_nextSrcSizeToDecompress` shouldn't be used in these cases. Furthermore, `ZSTD_nextSrcSizeToDecompress` returns a number which will be used as the objective, but if this number is greater than the number of bytes left in the ring buffer window, the code will enter a deadlock as its impossible to stream in more data than the current window allows for. We must wrap arround in these cases. During observation, this bug gets triggered once in a hundred launches on a mechanical drive and once every thousand launches on a solid-state drive.
The new fix checks if we decoded the entire frame before setting a new objective, while ensuring the new objective always fits within the given buffer window.