Common: fix major engine bug causing large performance impact
Only frustum culling was used instead of the full PVS, causing major CPU and GPU load on a bunch of area's. The game was still rendering the scene behind walls that obstruct the view completely. Forcing the cvar to 0 fixes this problem and gains over 100% FPS in many area's on lower end machines. This change also fixes a classic bug where large parts of the map dissapear on very specific locations and angles on the map