Recast: fix rare crash on degenerate polys in `closestPointOnPoly`
This code crashes when it encounters a degenerate polygon with no detail boundary edges. Very rare but it happened. Make sure we did find at least 1 detail polygon edge before we dereference `pmin` and `pmax` as they will be NULL otherwise. Slightly adjusted the usages of this API within the library to properly account for this edge case.