ModSystem: implement mod list mutex
the mod system will soon have a reload command, and the mod list will soon be used in FS_OpenAsyncFile to allow async files to be opened in mod search paths as well (this is required to make the engine read rpaks, starpaks, audio banks and color correction files from mod directories). FS_OpenAsyncFile however can be used by various threads so we have to bite the bullet and add a mutex in order for this to function correctly. The performance penalty caused by this mutex should be 0 since FS_OpenAsyncFile is not performance critical at all, unlike the FS read and polling functions, nor is the rest of the code which uses the mod list "hot".