Engine: fix edge case causing mod paks to load too early
If the VPK for the given level couldn't be loaded, this code will fire as the order the code gets run in the engine will change. So we actually should guard here for these cases to prevent this from crashing since a missing VPK is recoverable once the file problem is fixed.