FileSystem: allow mounting VPK's from mod directories
Previously you were only able to load VPK's from the MAIN directory of the game. The GameInfo.txt documentation on Valve largely applies to Titanfall and Apex as well, however not for VPK's; the VPK system has been modified in this engine and the search path logic was removed. This patch resolves the path to the mod directory making the VPK mounting from mod directories work again.