ModSystem: remove extraneous check for script compile list
m_bHasScriptCompileList will only be set during init, however if you are developing your mod you might be adding a compile list after the mod system has been initialized. Just look for it while we are loading mod scripts and let the file system decide whether we can load it or not, this happens during level load so this will have 0 impact on performance.