RTech: fix crash when loading decompressed RPak files
The tenary operator on the pak->maxCopySize member variable got optimized to always assign PAK_DECODE_OUT_RING_BUFFER_MASK. The result of the flag checks to determine if the pak is compressed has now been stored off in a stack variable which is reused. This now compiles to correct code and no longer crashes when loading. Also casted size values to 64 bit wide types to yield better output code.