Game: properly null-terminate format buffer on vsnprintf failure
vsnprintf always null terminates the array unless an error occured. However, we always null the last char in the array. If an error occured and we do not null the start of the array, we will print uninitialized memory (potentially until the end of the array). Check for error and null the start if encoding failed.