RTech: fix LoadLibrary() failure on paks loaded from mods
We need to look for the library from the directory the RPak was being loaded from. Use the new feature introduced in commit 17c07be6afb9a63f07d61f0014b7c94949660142 to make sure we always attempt to LoadLibrary() the image exactly from where the rpak was loaded from.