RTech: fix UAF in Pak_RequestSwap_f()
We cannot use members of the pak info structure after the unload was initiated as they get clobbered and freed. Store them on the stack until the unload job is finished and provide it back to Pak_LoadAsync(). This is a development only feature so this bug doesn't affect the rest of the game's runtime.